refactor: remove dead features and harden playback, backup, and QA foundations
- Remove Discord RPC (kizzy module), Listen Together, lastfm module, ForYouSuggestionEngine, and retired ListenBrainz manager
- Add backup preflight validation, safety restore archives, and transaction boundaries
- Add offline playback cache routing with completed-download detection
- Add search request gate preventing stale result publication
- Expand unit and instrumented test suites (playback policies, backup contracts, DB repair, session callback)
- Add QA runtime harness scripts and audit documentation

+++ app/src/main/kotlin/com/omnitune/app/ui/component/AudioVisualizer.kt
@@ -58,7 +58,7 @@ fun AudioVisualizer(
                 abs(sin(phase * 0.5f + i * 0.6f)) * 0.1f
             } else 0.05f
 
-            val barHeight = rawAmplitude.toFloat() * maxBarHeight
+            val barHeight = rawAmplitude * maxBarHeight
             val top = height / 2f - barHeight / 2f
             val bottom = height / 2f + barHeight / 2f
             val barColor = if (rawAmplitude > 0.5f) activeColor else color
